+++ doc/version_history.rst
@@ -7,6 +7,21 @@ Version History
 ###############
 
 ======
+v0.2.1
+======
+* Make sure that the MotionState enum always is used as a string instead of an int.
+* Make sure that the capacitor banks telemetry has the correct data types.
+* Spelling and punctuation corrections.
+* Split up exitFault to one command per subsystem.
+* Add infrastructure to request subsystem statuses regularly.
+* Fix race condition in telemetry code.
+
+Requires:
+
+* ts_tcpip
+* ts_utils
+
+======
 v0.2.0
 ======
 * Enable uploading documentation.